requirements.txt
Flask==3.1.0
numpy==1.26.4
scikit-learn==1.5.1
tensorflow==2.18.0
yfinance==0.2.54
pandas==2.2.2
gunicorn==23.0.0